git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@22725 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/web/forms/FormsService.java
@@ -378,7 +378,11 @@ public final class FormsService
             LOGGER.debug("removing configuration for " + formName +
                          " from web project " + this.nodeService.getProperty(parentRef, ContentModel.PROP_NAME));
          }
-         this.nodeService.removeChild(parentRef, ref);
+		 // ALF-3751: Validate this is the real form folder rather than a copy
+         if (childRef.getParentRef().equals(parentRef))
+         {
+            this.nodeService.removeChild(parentRef, ref);
+         }
       }
    }
